### 2020-07-18 #2485 @whedon
[PRE REVIEW]: NodePy: A package for the analysis of numerical ODE solvers
**Submitting author:** @ketch (<a href="http://orcid.org/0000-0002-1212-126X">David Ketcheson</a>)
**Repository:** <a href="https://github.com/ketch/nodepy" target ="_blank">https://github.com/ketch/nodepy</a>
**Version:** 0.9
