### 2020-07-18 #2485 @labarba
👋  @ketch -- have a look at `whedon`'s report above... we're going to need DOIs added to all references that have them. Go ahead and make those edits, then run the command `@whedon generate pdf` here, as a single-line comment.
