### 2020-07-19 07e1e0a
Add DOIs in JOSS paper bibliography.
