+++ paper.bib
@@ -49,6 +49,7 @@
 	Pages = {2113--2136},
 	Title = {Highly Efficient Strong Stability Preserving {R}unge-{K}utta Methods with Low-Storage Implementations},
 	Volume = {30},
+    doi = {10.1137/07070485X},
 	Year = {2008}}
 
 
@@ -58,8 +59,9 @@
 	Month = feb,
 	Number = {2},
 	Pages = {373--392},
-	Title = {{Optimal implicit strong stability preserving {R}unge--{K}utta methods}},
+	Title = {Optimal implicit strong stability preserving {R}unge--{K}utta methods},
 	Volume = {59},
+    doi = {10.1016/j.apnum.2008.03.034},
 	Year = {2009}}
 
 
@@ -70,6 +72,7 @@
 	Number = {5},
 	Pages = {1763--1773},
 	Title = {{R}unge--{K}utta methods with minimum storage implementations},
+    doi = {10.1016/j.jcp.2009.11.006},
 	Volume = {229},
 	Year = {2010}}
 
@@ -82,6 +85,7 @@
 	Pages = {1649},
 	Title = {{Step Sizes for Strong Stability Preservation with Downwind-Biased Operators}},
 	Volume = {49},
+    doi = {10.1137/100818674},
 	Year = {2011}}
 
 
@@ -93,6 +97,7 @@
 	Pages = {2618-2639},
 	Title = {Strong Stability Preserving Two-step {R}unge--{K}utta Methods},
 	Volume = {49},
+    doi = {10.1137/10080960X},
 	Year = {2011}}
 
 
@@ -103,6 +108,7 @@
 	Pages = {2149--2165},
 	Title = {Strong Stability Preserving Explicit {R}unge--{K}utta Methods of Maximal Effective Order},
 	Volume = {51},
+    doi = {10.1137/120884201},
 	Year = {2013}}
 
 @article{2014_internal_stability,
@@ -112,6 +118,7 @@
 	Pages = {2227--2249},
 	Title = {Internal error propagation in explicit {R}unge--{K}utta methods},
 	Volume = {52},
+    doi = {10.1137/130936245},
 	Year = {2014}}
 
 
@@ -122,6 +129,7 @@
 	Pages = {175-200},
 	Title = {A comparison of high order explicit {R}unge-{K}utta, extrapolation, and deferred correction methods in serial and parallel},
 	Volume = {9},
+    doi = {10.2140/camcos.2014.9.175},
 	Year = {2014}}
 
 
@@ -132,6 +140,7 @@
 	Pages = {2799-2832},
 	Title = {Strong stability preserving explicit linear multistep methods with variable step size},
 	Volume = {54},
+    doi = {10.1137/15M101717X},
 	Year = {2016}}
 
 
@@ -142,6 +151,7 @@
 	Pages = {944--958},
 	Title = {Dense output for strong stability preserving {R}unge--{K}utta methods},
 	Volume = {71},
+    doi = {10.1007/s10915-016-0331-5},
 	Year = {2017}}
 
 
@@ -151,6 +161,7 @@
 	Pages = {747-769},
 	Title = {Strong Stability Preserving Multistep {R}unge-{K}utta Methods},
 	Volume = {86},
+    doi = {10.1090/mcom/3115},
 	Year = {2017}}
 
 
@@ -161,6 +172,7 @@
 	Pages = {1337--1369},
 	Title = {Optimal Monotonicity-Preserving Perturbations of a Given {R}unge--{K}utta Method},
 	Volume = {76},
+    doi = {10.1007/s10915-018-0664-3},
 	Year = {2018}}
 
 
@@ -171,6 +183,7 @@
[…]
