### 2020-07-19 #2485 @ketch
I've added DOIs for everything that has them.  Note that one of the DOIs suggested by the bot actually corresponds to a book review rather than the book, so I didn't add it.
