### 2020-07-24 #2515 @whedon
[REVIEW]: NodePy: A package for the analysis of numerical ODE solvers
**Submitting author:** @ketch (<a href="http://orcid.org/0000-0002-1212-126X">David Ketcheson</a>)
**Repository:** <a href="https://github.com/ketch/nodepy" target ="_blank">https://github.com/ketch/nodepy</a>
**Version:** 1.0.1
