### 2020-08-19 ketch/nodepy#83 @fruzsinaagocs
Convergence test error in introductory example
When running [Introduction to NodePy.ipynb](https://github.com/ketch/nodepy/blob/master/examples/Introduction%20to%20NodePy.ipynb), I get the following Sympy warning in the Convergence test section:
```bash
/home/fruzsina/all-virtual-envs/nodepy-env/lib/python3.8/site-packages/sympy/__init__.py:672: SymPyDeprecationWarning:

importing sympy.core.basic with 'from sympy import *' has been
deprecated since SymPy 1.6. Use import sympy.core.basic instead. See
https://github.com/sympy/sympy/issues/18245 for more info.

  self.Warn(
```

And the output is showing that the error increases with function evaluations:
![output-nodepy-1](https://user-images.githubusercontent.com/19972401/90685080-2ba28e00-e261-11ea-8dde-da47de25b033.png)

The same plot in the static version of the notebook (as seen on GitHub) seems correct:
![nodepy-output-2](https://user-images.githubusercontent.com/19972401/90685336-8c31cb00-e261-11ea-9beb-a4cf9d5638dc.png)
