### 2020-08-19 ketch/nodepy#84 @fruzsinaagocs
Sympy error in Method of lines analysis example
In the example notebook titled [Method of lines analysis](https://github.com/ketch/nodepy/blob/master/examples/Method%20of%20lines%20analysis.ipynb), running the cell

```python
rkc = rk.RKC2(10)
rk.linearly_stable_step_size(rkc,centered_diff)
```
gives
```bash
<ipython-input-7-6a581f261783> in <module>
----> 1 rkc = rk.RKC2(10)
      2 rk.linearly_stable_step_size(rkc,centered_diff)

~/all-virtual-envs/nodepy-env/lib/python3.8/site-packages/nodepy/runge_kutta_method.py in RKC2(m, epsilon)
   3504
   3505     x=sympy.Symbol('x')
-> 3506     Tm=sympy.special.polynomials.chebyshevt_poly(m,x)
   3507
   3508     w0=one+sympy.Rational(epsilon,m**2)

AttributeError: module 'sympy' has no attribute 'special'
```
I think the relevant function implementing Chebyshev polynomials moved to `sympy.functions.special.polynomials.chebyshevt` ([sympy docs](https://docs.sympy.org/latest/modules/functions/special.html#chebyshev-polynomials)).

This notebook also relies on `scipy`, which is not a listed dependency.
