### 2020-08-19 ketch/nodepy#85 @fruzsinaagocs
Old time module used in stability polynomials example
In the example titled [Speed of symbolic computation of stability polynomials](https://github.com/ketch/nodepy/blob/master/examples/Stability%20functions%20and%20order%20stars.ipynb), cell 4
```python
time_stability_polynomial_calculation(rk.extrap,pmax=7)
```
gives
```bash
<ipython-input-4-168ca6611554> in <module>
----> 1 time_stability_polynomial_calculation(rk.extrap,pmax=7)

<ipython-input-3-0be4a33ec95c> in time_stability_polynomial_calculation(method, pmin, pmax, approaches, SO)
     10         myrk = method(order)
     11         for j, approach in enumerate(approaches):
---> 12             p,q,t = approach(myrk,SO=SO)
     13             times[j][i] = t
     14

<ipython-input-2-a436ffccb2fc> in lower_triangular_solve(rk, SO)
      1 def lower_triangular_solve(rk,SO=False): # Using lower_triangular_solve, which is always faster than LUsolve
----> 2     start = time.clock()
      3     z=sympy.var('z')
      4     I = sympy.matrices.eye(len(rk))
      5     if SO:

AttributeError: module 'time' has no attribute 'clock'

```
This is due to [updates](https://docs.python.org/3/whatsnew/3.8.html#api-and-feature-removals) to to the `time` module since Python 3.3:

> The function time.clock() has been removed, after having been deprecated since Python 3.3: use time.perf_counter() or time.process_time() instead, depending on your requirements, to have well-defined behavior. (Contributed by Matthias Bussonnier in bpo-36895.)
