### 2020-08-19 ketch/nodepy#86 @fruzsinaagocs
Undefined quantity in very high order methods example
This [example](https://github.com/ketch/nodepy/blob/master/examples/Very%20high%20order%20methods.ipynb) is great! It would be even better with some additional comments. There is, in additions, an undefined quantity `ex3` in cell 6:

```python
ex3.num_seq_dep_stages()
```
