### 2020-08-20 2d3b0f3
fix #86 (typo in example notebook for DeC and extrapolation
