### 2020-08-20 47a24ba
fix #86 (typo in example notebook for DeC and extrapolation
