### 2020-08-20 3513aa1
fix #85 (stability polynomial speed example)
