### 2020-08-20 c6cd06a
fix #85 (stability polynomial speed example)
