### 2020-08-20 ketch/nodepy#87 @ranocha
Run example notebooks as tests automatically
Is there a way to test all example notebooks automatically? This would have prevented the recent issues #83, #84, #85, #86.
