### 2020-08-20 ketch/nodepy#84 @ranocha
Thanks for the report. The Chebyshev polynomials are already fixed in c4787f17a8fa36e62716b2d294c3617b25455fd4 - there should just be a new tag.

@ketch: How would you like to handle the dependency issue?
