### 2020-08-20 ketch/nodepy#83 @ranocha
The SymPy warning might have been fixed already on `master` (at least I don't get this warning anymore). I'll push a fix for the other issue.
