### 2020-08-20 8e6c62f
fix #83 (wrong binding of u0 in lambda functions for ivp)
