### 2020-08-20 9d469e6
fix #83 (wrong binding of u0 in lambda functions for ivp)
