### 2020-08-20 ketch/nodepy#87 @ketch
Yes. I have set that up for other projects and will do the same here.
