### 2020-08-20 #2515 @fruzsinaagocs
Hi all,

Thanks for writing this great package! Bringing together techniques from
classic numerical analysis works was much needed, and this certainly has the potential to be
widely used in research. I've ticked off most things on the reviewer checklist, I'll
organise the remaining points below.

### Python 3 compatibility and dependencies

* In the readthedocs [Dependencies](https://nodepy.readthedocs.io/en/latest/index.html#dependencies) section, it is stated that NodePy depends on (among others):
    * Python 2.7 (current, stable version: 3.8)
    * Sympy 0.7 (current, stable version: 1.6)
Python 2.7 is unsupported as of this year, and Sympy has deprecated some of the
functions used in the package since. Ideally, the package should support Python
3, but at least should make it very clear which version of the dependencies it
requires in the `README.md` and docs alike, and fetch the intended versions of
the dependencies. I can see that Travis CI is already testing in Python 3.5 (as
well as 2.7), which suggests that this work was planned/is in progress.
* There is a mismatch between what optional dependencies the package has:
  `README.md` lists cvxpy, the docs mention networkx. It should be clear which
  modules rely on these.
* Some example notebooks make use of scipy, this should be added as a dependency
  and fetched.

### Installation

`pip install` and installation from source were both quick and easy. Just some
minor points:

* The docs [Quick Start
  Guide](https://nodepy.readthedocs.io/en/latest/quickstart.html#obtaining-nodepy)
  says that the pip version is often outdated and the package should be
  installed from source, whereas `README.md` suggets `pip install`. These should
  mathc, and ideally the pip version should be up to date.
* The same guide suggests that the user updates the `PYTHONPATH` variable as
```bash
export PYTHONPATH=/user/home/python/
```
but this overwrites `PYTHONPATH`. I would change this to
```bash
export PYTHONPATH=$PYTHONPATH:/user/home/python/
```
so that it is appended to instead.

### Documentation

* A link to the Useg Guide from the docs [Examples section](https://nodepy.readthedocs.io/en/latest/quickstart.html#examples) is broken. This would be a useful feature.
* An alternative docs website link from the docs [Documentation section](https://nodepy.readthedocs.io/en/latest/quickstart.html#nodepy-documentation) is also broken.
* There is a note about user guide being incomplete and not up-to-date [here in the docs](https://nodepy.readthedocs.io/en/latest/index.html). It's not clear which guide is being referred to.
* Throughout the docs, the rendering of LaTeX equations is broken, e.g. [here](https://nodepy.readthedocs.io/en/latest/rkm.html#runge-kutta-methods).
* In general, the `README.md` and docs should be more synchronised in their content.

### Examples

* scipy dependency should be added, as mentioned above.
* having worked though the examples, I submittted issues https://github.com/ketch/nodepy/issues/83,https://github.com/ketch/nodepy/issues/84, https://github.com/ketch/nodepy/issues/85, https://github.com/ketch/nodepy/issues/86 https://github.com/ketch/nodepy/issues/86 mainly to do with Python 3/Sympy 1.6 compatibility.
* It would be nice to run the example notebooks automatically, as mentioned
  already in https://github.com/ketch/nodepy/issues/87
* The example notebooks are really neat. They would be even better with some
  more comments and labelling of figure axes.

### Community guidelines

* Guidelines on how to seek support should be added.

### Testing

Testing's done via Travis continuous integration, the unit tests are found in
[`nodepy/unit_tests.py`](https://github.com/ketch/nodepy/blob/master/nodepy/unit_tests.py). Both in the comments and in the docs, it is mentioned that the unit tests are to be updated, which I support. An example of an incomplete test case is the `LinearMultistepTestCase`.

### Miscellaneous

* The [paper](https://github.com/openjournals/joss-papers/blob/joss.02515/joss.02515/10.21105.joss.02515.pdf) lists numerical methods the package has object-oriented representations for, one being dense output formulas. This would be incredibly useful, but I couldn't find anything about the topic in the docs, and the module reference includes some `TODO`s in the [relevant docstring](https://nodepy.readthedocs.io/en/latest/modules/runge_kutta_method.html#nodepy.runge_kutta_method.ContinuousRungeKuttaMethod). This should be updated, if such feature is planned.
