### 2020-08-21 07624bf
update remaining examples to new version of python
