### 2020-08-23 1ba5197
Improvements based on reviewer comments.
Add narrative comments to some of the example notebooks.
Load Mathjax over https so hopefully it renders on readthedocs.io.
Remove outdated installation instructions.
State dependencies consistently.
