+++ requirements.txt
@@ -1,4 +1,3 @@
 numpy>=1.6
-sympy>=0.7.6
-matplotlib>=1.2
-networkx>=2.0
+sympy>=1.6
+matplotlib>=3.0
