### 2020-08-23 5ff088d
Improvements based on reviewer comments.
Add narrative comments to some of the example notebooks.
Load Mathjax over https so hopefully it renders on readthedocs.io.
Remove outdated installation instructions.
State dependencies consistently.
