+++ doc/conf.py
@@ -40,7 +40,7 @@ extensions = [
               'sphinxcontrib.bibtex']
 
 # Point to the MathJax CDN (latest MathJax.js stable release)
-mathjax_path = 'http://cdn.mathjax.org/mathjax/latest/MathJax.js?config=TeX-AMS-MML_HTMLorMML'
+mathjax_path = 'https://cdn.mathjax.org/mathjax/latest/MathJax.js?config=TeX-AMS-MML_HTMLorMML'
 
 # Add any paths that contain templates here, relative to this directory.
 templates_path = ['_templates']
