### 2020-08-23 5ed03f8
Automatically test notebooks.
Closes #87.
Had to omit a couple of notebooks because they run for a very
long time, which can cause timeout errors with nbconvert or Travis.
