+++ requirements.txt
@@ -1,3 +1,3 @@
 numpy>=1.6
-sympy>=1.6
+sympy>=1.5
 matplotlib>=3.0
