### 2020-08-23 aad22f4
Install jupyter, nbformat, and nbconvert on Travis.
