### 2020-08-23 4e82e22
Install jupyter, nbformat, and nbconvert on Travis.
