### 2020-08-23 3354e96
Stop testing for Python 2.7.
