### 2020-08-23 cef1e4b
Drop Python 2.7 support.
