### 2020-08-23 b6cff1f
Bump version to 1.0.0.
