### 2020-08-27 ketch/nodepy#84 @ketch
Scipy is now listed as an optional dependency (fixed in 5ff088df09a77f13bdac1ff49bd846bff62447ca).
