### 2020-09-07 #2515 @fruzsinaagocs
These changes look great, @ketch.

I have now ran the example notebooks, they run smoothly and have a good amount of detail. Running them automatically as tests is a neat idea.

Documenting a package with so much functionality is no small task, but with the existing examples and docs I trust that the few missing bits will soon be covered and I'm happy to tick this off. (As it happens, testing dense output from an arbitrary RK method is of particular interest to my research!)

I recommend the package for publication.
