### 2020-09-09 #2515 @ketch
@fruzsinaagocs Thanks again for the detailed and thorough review.  If you let me know more specifically what you're hoping to test regarding dense output, I can probably add the needed functionality or documentation to the package.
