### 2020-09-09 #2515 @fruzsinaagocs
> @fruzsinaagocs Thanks again for the detailed and thorough review. If you let me know more specifically what you're hoping to test regarding dense output, I can probably add the needed functionality or documentation to the package.

I'm specifically looking for testing the efficiency of the continuous extension of Runge-Kutta formulas. To compare efficiencies of two formulas, [Shampine: Some practical Runge-Kutta formulas](https://doi.org/10.2307/2008219 ) offers a method in Section 3. "Interpolation", based on the truncation error coefficients of the formula, expressions for which Chapter 6 of [Hall, Watt: Modern numerical methods for ODEs](https://doi.org/10.1017/S0013091500026699) is cited. If `Nodepy` could compute this efficiency metric for an arbitrary continuous formula, that would be very helpful.

Going one step further, if one could automatically compute the coefficients for a continuous extension of a given RK formula by solving the order conditions from e.g. [Horn: Fourth- and fifth-order, scaled Runge-Kutta algorithms for treating dense output](https://www.jstor.org/stable/2157271) and determine at which points one could get a 'free' or 'cheap' output of order p, it would make the construction of dense output formulas a lot easier!

Does any of the above sound like it could be added as a feature?
