### 2020-10-02 #2515 @VivianePons
Hi @wingtorres have you had time to look at it?
