### 2020-10-23 #2515 @VivianePons
I had to remove Lisandro as reviewer due our [conflict of interest policy](https://github.com/openjournals/joss/blob/master/COI.md) I am looking for a replacement
