### 2020-11-03 20daabf
Update Travis CI badge; closes #91
