### 2020-11-09 #2515 @highlando
Alright. I can say that this looks like an advanced and useful piece of software. The formal requirements are all satisfied. I don't have any major issues. Only a few minor things:
- [ ] The link to documentation on http://numerics.kaust.edu.sa/nodepy is forwarded to *readthedocs*. Maybe have only one repo and one address for the documentation.
- [ ] Examples -- maybe one on how to apply these analyses for custom methods. (Although from the class descriptions it is very clear how this may work).
- [ ] The community support/contribution is very brief. Maybe add some more words.
- [ ] The paper does not refer to similar software or other software that might be helpful with these tasks. If there is none, the authors may just state this fact.
