### 2020-11-10 #2515 @ketch
Thanks @highlando.

> Maybe have only one repo and one address for the documentation.

Good idea.  I've tried to get rid of all the old links, but left the redirect there in case there are some still floating around the internet.  The KAUST Numerics site is in a major transition at the moment, but when that is finished those links will be updated too.

> Examples -- maybe one on how to apply these analyses for custom methods. (Although from the class descriptions it is very clear how this may work).

Good idea; I've added this in the "Introduction to Nodepy" notebook in examples/.

> The community support/contribution is very brief. Maybe add some more words.

In my opinion, length itself is not a virtue.  But is there specific information that you think we should add?  And if so, are you referring to the README, the docs, or the paper?

> The paper does not refer to similar software or other software that might be helpful with these tasks. If there is none, the authors may just state this fact.

I've added a sentence to this effect at the beginning of the section.
