### 2020-11-10 d10cfa0
add JOSS reference to RK-Opt
