### 2020-11-10 #2515 @highlando
Hi @ketch

I'm OK with the extra link -- just a warning that this extra work when maintaining.

Thanks for the extra line in the *Introduction to Nodepy* notebook. Much appreciated.

As for the *support/contribution*. I was referring to the [documentation](https://nodepy.readthedocs.io/en/latest/about.html#contributing). Because of its brevity it is not so prominent. Maybe add a link to the github repo and also offer to open an issue in the issue tracker there. But I'm totally fine with its current state.

Thanks for the additional information that there is no similar software around.

I'm completely satisfied with the submission, @VivianePons
