### 2020-11-11 70eaf60
Update "about" page, expanding "contributing" section.
