### 2020-11-11 #2515 @ketch
Thanks for calling my attention to that page in the docs; it was outdated in other ways, too.  I've expanded the contributing section with a couple of links for Github issues and PRs.
