### 2020-11-12 #2515 @VivianePons
Hi @highlando the box "State of the field" is not checked in the issue description, if you are now satisfied, can you tick it?
