### 2020-11-16 27070f4
Bump version to 1.0.1 for Zenodo release.
