+++ doc/conf.py
@@ -68,7 +68,7 @@ copyright = u'2009-2020, David Ketcheson'
 # The short X.Y version.
 version = '1.0'
 # The full version, including alpha/beta/rc tags.
-release = '1.0.0'
+release = '1.0.1'
 
 # The language for content autogenerated by Sphinx. Refer to documentation
 # for a list of supported languages.
