### 2020-11-16 #2515 @ketch
@VivianePons Thanks to you and the reviewers for all your hard work!  I've made a release on Zenodo and I think everything is in order.
